% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{Get game summary in a dataframe}
\usage{
game(x)
}
\arguments{
\item{x}{a quackle or cross-tables game number or link, or file path}
}
\value{
data.frame game summary
}
\description{
Get game summary in a dataframe
}
\examples{
game(game_file(28283))
game(game_file("http://www.cross-tables.com/annotated/selfgcg/281/anno28152.gcg"))
game(game_file("http://www.cross-tables.com/annotated.php?u=28190#0#"))
}
