% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_all_words.R
\name{game_all_words}
\alias{game_all_words}
\title{Get all words on board in a game}
\usage{
game_all_words(df)
}
\arguments{
\item{x}{a dataframe produced by game() function}
}
\value{
a vector of words on the board in a game
}
\description{
Get all words on board in a game
}
\examples{
game_words(game(game_file(28283)))
}
