% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_board.R
\name{game_board}
\alias{game_board}
\title{Get final game board of a game}
\usage{
game_board(x, type = "pretty")
}
\arguments{
\item{x}{a dataframe produced by game() function}
}
\value{
matrix of game board
}
\description{
Get final game board of a game
}
\examples{
game_board(game(game_file(28283)))
game_board(game(game_file(28283)), type="clean")
game_board(game(game_file(28283)), type="dots")
}
