% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_board_geom.R
\name{game_board_geom}
\alias{game_board_geom}
\title{Get final game board of a game in ggplot2 style}
\usage{
game_board_geom(x, fsize = 5, asize = 1, gsize = 0.8, facetype = "bold",
  alp = 0.4, turn = NULL)
}
\arguments{
\item{x}{a dataframe produced by game() function}

\item{fsize}{font size for text}

\item{asize}{font size for axis labels}

\item{gsize}{grid size}

\item{facetype}{font type}

\item{alp}{transparency of tile colors}
}
\value{
ggplot2 plot of game board
}
\description{
Get final game board of a game in ggplot2 style
}
\examples{
game_board_geom(game(game_file(28283)))
}
