% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_file.R
\name{game_file}
\alias{game_file}
\title{Get raw game text data from a quackle or cross-tables game number or link or filepath.}
\usage{
game_file(v)
}
\arguments{
\item{x}{a quackle or cross-tables game number or link, or file path}
}
\value{
data.frame game summary
}
\description{
Get raw game text data from a quackle or cross-tables game number or link or filepath.
}
\examples{
game_file(28283)
game_file("http://www.cross-tables.com/annotated/selfgcg/281/anno28152.gcg")
game_file("http://www.cross-tables.com/annotated.php?u=28190#0#")
}
