% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_summary.R
\name{game_summary}
\alias{game_summary}
\title{Get summary statistics of a game in a dataframe}
\usage{
game_summary(df)
}
\arguments{
\item{x}{a dataframe produced by game() function}
}
\value{
data.frame game summary statistics
}
\description{
Get summary statistics of a game in a dataframe
}
\examples{
game_summary(game(game_file(28283)))
}
