% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_words.R
\name{game_words}
\alias{game_words}
\title{Get all words played in a game}
\usage{
game_words(x)
}
\arguments{
\item{x}{a dataframe produced by game() function}
}
\value{
a vector of words played by each player
}
\description{
Get all words played in a game
}
\examples{
game_words(game(game_file(28283)))
}
