% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_dist.R
\name{tile_dist}
\alias{tile_dist}
\title{Get tile distribtuion of a game}
\usage{
tile_dist(df)
}
\arguments{
\item{x}{a dataframe produced by game() function}
}
\value{
table of tile distributions played
}
\description{
Get tile distribtuion of a game
}
\examples{
tile_dist(game(game_file(28283)))
}
